# Conflicts:
#	GCFG/Gopal Chan's Fun Game/Assets/Adarsh/anim/human.controller
#	GCFG/Gopal Chan's Fun Game/Assets/Akhil/Animation/Sword.meta
#	GCFG/Gopal Chan's Fun Game/Assets/Akhil/Prefabs/Resources/Tower.prefab

+++ GCFG/Gopal Chan's Fun Game/Assets/Gopal/Weapons/Scripts/Bullet.cs
@@ -8,7 +8,7 @@ namespace Gopal
     {
         [SerializeField]
         private float speed = 5f;
-        public int damage { get; set; }
+        public float damage { get; set; }
 
         private void OnTriggerEnter(Collider other)
         {
